changed MAKEJWT to JWT.py, added extract_id function (tested and working), moved decodeJWT code to JWT.py from Authorization filters
